+++ inst/sql/CreateKeeperCohort.sql
@@ -1,3 +1,5 @@
+DROP TABLE IF EXISTS #cohort;
+
 SELECT subject_id,
 	generated_id,
 	cohort_start_date
@@ -9,7 +11,7 @@ FROM (
 	FROM @cohort_table
 	WHERE cohort_definition_id = @cohort_definition_id
 {@use_person_ids} ? {		AND subject_id IN (SELECT CAST(person_id AS BIGINT) FROM #person_ids) }
-)
+) tmp
 WHERE generated_id <= @sample_size;